// src/types/FeedVideoType.ts
export type FeedVideoType = {
  kind: string;
  etag: string;
  id: string;
  snippet: Snippet;
  contentDetails: ContentDetails;
  statistics: Statistics;
};

type Statistics = {
  viewCount: string;
  likeCount: string;
  favoriteCount: string;
  commentCount: string;
};

type ContentDetails = {
  duration: string;
  dimension: string;
  definition: string;
  caption: string;
  licensedContent: boolean;
  contentRating: Record<string, string>;
  projection: string;
};

type Snippet = {
  publishedAt: string;
  channelId: string;
  title: string;
  description: string;
  thumbnails: Thumbnails;
  channelTitle: string;
  categoryId: string;
  liveBroadcastContent: string;
  defaultLanguage: string;
  localized: Localized;
  defaultAudioLanguage: string;
};

type Localized = {
  title: string;
  description: string;
};

type Thumbnails = {
  default: Default;
  medium: Default;
  high: Default;
  standard: Default;
  maxres: Default;
};

type Default = {
  url: string;
  width: number;
  height: number;
};